requirements.txt
flask
python-dotenv
pyjwt